requirements.txt
django
djangorestframework
markdown
django-filter
aiogram
python-dotenv
